Remember Me in Login

+++ src/main/java/art/school/statik/MonthForStatistik.java
@@ -3,6 +3,7 @@ package art.school.statik;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.Month;
 import java.util.List;
 
@@ -12,6 +13,6 @@ public class MonthForStatistik {
 
     private String name;
     private Month monat;
-    private Double value;
+    private BigDecimal value;
     private List<WeeksForStatistik> childrens;
 }
