+++ src/main/java/art/school/statik/WeeksForStatistik.java
@@ -3,6 +3,7 @@ package art.school.statik;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -11,10 +12,10 @@ public class WeeksForStatistik {
 
     private Integer nummer;
     private String name;
-    private Double value;
+    private BigDecimal value;
     private List<DaysForStatistik> childrens;
 
-    public WeeksForStatistik(Integer nummer, String name, Double value) {
+    public WeeksForStatistik(Integer nummer, String name, BigDecimal value) {
         this.nummer = nummer;
         this.name = name;
         this.value = value;
